Separate RegionViewer and region reducer into different packages
