+++ packages/region-viewer/src/example/index.js
@@ -1,0 +1,1 @@
+export { default } from './RegionViewerExample'
