+++ packages/region-viewer/src/index.js
@@ -1,0 +1,5 @@
+export { attributeConfig, coverageConfigClassic, coverageConfigNew } from './regionViewerStyles'
+
+export { RegionViewer } from './RegionViewer'
+
+export { default as GeneViewer } from './ConnectedGeneViewer'
