+++ packages/region/package.json
@@ -7,30 +7,15 @@
     "lib"
   ],
   "dependencies": {
-    "@broad/redux-genes": "*",
     "@broad/redux-variants": "*",
     "@broad/ui": "*",
-    "@broad/utilities": "*",
-    "react-apollo": "^2.0.1",
-    "react-redux": "^5.0.6"
-  },
-  "devDependencies": {
-    "@broad/track-position-table": "*",
-    "@broad/track-transcript": "*",
-    "apollo-cache-inmemory": "^1.1.1",
-    "apollo-client": "^2.0.3",
-    "apollo-link-http": "^1.2.0",
-    "graphql-tag": "^2.5.0",
     "immutable": "^4.0.0-rc.9",
     "keymirror": "^0.1.1",
+    "prop-types": "^15.5.10",
     "react-redux": "^5.0.6",
-    "redux": "3.6.0",
-    "redux-thunk": "^2.2.0",
     "reselect": "^3.0.1"
   },
   "peerDependencies": {
-    "prop-types": "^15.5.10",
-    "react": "^15.5.4",
-    "styled-components": "^2.1.2"
+    "react": "^15.5.4"
   }
 }
