--- packages/region/src/example/index.js
@@ -1,4 +1,0 @@
-import Example from './apollo.example'
-// import Example from './redux.example'
-
-export default Example
