+++ packages/region/src/index.js
@@ -10,18 +10,4 @@ export {
   currentChromosome,
 } from './regionRedux'
 
-export {
-  coverageConfigClassic,
-  coverageConfigNew,
-  attributeConfig,
-} from './regionViewerStyles'
-
 export { RegionHoc } from './RegionHoc'
-
-export {
-  default as RegionViewer,
-} from './RegionViewer'
-
-export {
-  default as GeneViewer,
-} from './GeneViewerReduxConnected'
