--- packages/region/src/query.graphql
@@ -1,14 +1,0 @@
-query GeneViewer($geneName: String!) {
-  gene(gene_name: $geneName) {
-    id: gene_name
-    gene_name
-    transcript {
-      exons {
-        feature_type
-        start
-        stop
-        strand
-      }
-    }
-  }
-}
