+++ packages/track-stacked-bar/src/example/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled from 'styled-components'
-import RegionViewer from '@broad/region'
+import { RegionViewer } from '@broad/region-viewer'
 
 import {
   // coverageConfigClassic,
